telis_recruitment/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-ratelimit>=4.1.0
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6
sib-api-v3-sdk>=7.6.0
psutil>=5.9.0
openpyxl>=3.1.0
django-unfold>=0.47.0
